removed old reducer
