+++ frontend/src/store/current_interval/crew_involvements.ts
@@ -1,4 +1,6 @@
-import type { CrewInvolvement } from "../../api/Api";
+import type { CrewInvolvement, Person } from "../../api/Api";
+import { compareStrings } from "../../utilities/comparison";
+import type { PeopleObjectMap } from "../people";
 
 export function forCrew(involvements: CrewInvolvement[], crewId: number): CrewInvolvement[] {
   return involvements.filter((involvement) => involvement.crew_id === crewId);
@@ -6,6 +8,13 @@ export function forCrew(involvements: CrewInvolvement[], crewId: number): CrewIn
 
 export function forPerson(involvements: CrewInvolvement[], personId: number): CrewInvolvement[] {
   return involvements.filter((involvement) => involvement.person_id === personId);
+}
+
+export function asPeopleAlphaSorted<T extends { person_id: number }>(involvements: T[], people: PeopleObjectMap): Person[] {
+  return involvements
+    .map((involvement) => people[involvement.person_id])
+    .filter(Boolean)
+    .sort(compareStrings("display_name"));
 }
 
 export function updatePersonCrewInvolvements(existing: CrewInvolvement[], newInvolvements: CrewInvolvement[], personId: number): CrewInvolvement[] {
